refactor(feeder): move doRequest and Validatable to a new request.go file
